Move TestDataflowRunner into dataflow package
